회원가입 API 오류 수정. 회원가입 API 완료
