--- alembic/versions/07ee0dfaa23c_아티클_엔티티에_문제_번호_어트리뷰트_추가_문제_번호로_검색_api까지.py
@@ -1,28 +1,0 @@
-"""아티클 엔티티에 문제 번호 어트리뷰트 추가, 문제 번호로 검색 API까지
-
-Revision ID: 07ee0dfaa23c
-Revises: 99ec830f2cf9
-Create Date: 2025-03-03 15:15:30.135582
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '07ee0dfaa23c'
-down_revision = '99ec830f2cf9'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('Article', sa.Column('problem_numbers', sa.JSON(), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('Article', 'problem_numbers')
-    # ### end Alembic commands ###
