+++ alembic/versions/3901996413e2_initial_migration.py
@@ -1,0 +1,47 @@
+"""initial migration
+
+Revision ID: 3901996413e2
+Revises: 
+Create Date: 2025-05-10 19:54:45.822993
+
+"""
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision = '3901996413e2'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+
+def upgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('blocked_tokens',
+    sa.Column('token_id', sa.String(length=255), nullable=False),
+    sa.Column('expired_at', sa.DateTime(timezone=True), nullable=False),
+    sa.PrimaryKeyConstraint('token_id')
+    )
+    op.create_table('user',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('username', sa.String(length=20), nullable=True),
+    sa.Column('login_id', sa.String(length=255), nullable=False),
+    sa.Column('password', sa.String(length=100), nullable=False),
+    sa.Column('birth', sa.DateTime(timezone=True), nullable=True),
+    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_index(op.f('ix_user_login_id'), 'user', ['login_id'], unique=True)
+    op.create_index(op.f('ix_user_username'), 'user', ['username'], unique=True)
+    # ### end Alembic commands ###
+
+
+def downgrade():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_index(op.f('ix_user_username'), table_name='user')
+    op.drop_index(op.f('ix_user_login_id'), table_name='user')
+    op.drop_table('user')
+    op.drop_table('blocked_tokens')
+    # ### end Alembic commands ###
