--- alembic/versions/8d7ee3556a55_initial_migration.py
@@ -1,201 +1,0 @@
-"""initial migration
-
-Revision ID: 8d7ee3556a55
-Revises: 07ee0dfaa23c
-Create Date: 2025-05-10 18:34:22.670684
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = '8d7ee3556a55'
-down_revision = '07ee0dfaa23c'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('blocked_tokens',
-    sa.Column('token_id', sa.String(length=255), nullable=False),
-    sa.Column('expired_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('token_id')
-    )
-    op.create_table('user',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('username', sa.String(length=20), nullable=True),
-    sa.Column('email', sa.String(length=255), nullable=False),
-    sa.Column('password', sa.String(length=100), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_user_email'), 'user', ['email'], unique=True)
-    op.create_index(op.f('ix_user_username'), 'user', ['username'], unique=True)
-    op.create_table('blog',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('blog_name', sa.String(length=100), nullable=False),
-    sa.Column('description', sa.String(length=255), nullable=True),
-    sa.Column('main_image_url', sa.String(length=255), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('default_category_id', sa.Integer(), nullable=True),
-    sa.Column('user_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id', name='unique_user_blog')
-    )
-    op.create_table('message',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('is_read', sa.Boolean(), nullable=False),
-    sa.Column('sender_id', sa.BigInteger(), nullable=False),
-    sa.Column('recipient_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['recipient_id'], ['user.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['sender_id'], ['user.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('blog_subscription',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('subscriber_id', sa.BigInteger(), nullable=True),
-    sa.Column('subscribed_id', sa.BigInteger(), nullable=True),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.ForeignKeyConstraint(['subscribed_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['subscriber_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('subscriber_id', 'subscribed_id', name='unique_subscription')
-    )
-    op.create_table('category',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('name', sa.String(length=50), nullable=False),
-    sa.Column('level', sa.Integer(), nullable=False),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.Column('parent_id', sa.BigInteger(), nullable=True),
-    sa.CheckConstraint('level IN (1, 2)', name='valid_category_level'),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ),
-    sa.ForeignKeyConstraint(['parent_id'], ['category.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_category_name'), 'category', ['name'], unique=False)
-    op.create_table('draft',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('title', sa.String(length=80), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_draft_title'), 'draft', ['title'], unique=False)
-    op.create_table('Article',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('secret', sa.Integer(), nullable=False),
-    sa.Column('protected', sa.Integer(), nullable=False),
-    sa.Column('password', sa.String(length=60), nullable=True),
-    sa.Column('comments_enabled', sa.Integer(), nullable=False),
-    sa.Column('title', sa.String(length=80), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('description', sa.Text(), nullable=False),
-    sa.Column('views', sa.Integer(), nullable=False),
-    sa.Column('main_image_url', sa.String(length=255), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.Column('problem_numbers', sa.JSON(), nullable=True),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.Column('category_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['category_id'], ['category.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_Article_title'), 'Article', ['title'], unique=False)
-    op.create_table('blog_like',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.Column('article_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['article_id'], ['Article.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('comment',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('content', sa.String(length=500), nullable=False),
-    sa.Column('level', sa.Integer(), nullable=False),
-    sa.Column('secret', sa.Integer(), nullable=False),
-    sa.Column('user_name', sa.String(length=20), nullable=False),
-    sa.Column('user_id', sa.BigInteger(), nullable=True),
-    sa.Column('user_blog_id', sa.Integer(), nullable=False),
-    sa.Column('article_id', sa.BigInteger(), nullable=True),
-    sa.Column('blog_id', sa.BigInteger(), nullable=True),
-    sa.Column('parent_id', sa.BigInteger(), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), nullable=False),
-    sa.CheckConstraint('(article_id IS NOT NULL OR blog_id IS NOT NULL)', name='check_article_or_blog'),
-    sa.ForeignKeyConstraint(['article_id'], ['Article.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['parent_id'], ['comment.id'], ),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_comment_article_id'), 'comment', ['article_id'], unique=False)
-    op.create_index(op.f('ix_comment_blog_id'), 'comment', ['blog_id'], unique=False)
-    op.create_table('image',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('file_url', sa.String(length=200), nullable=False),
-    sa.Column('article_id', sa.BigInteger(), nullable=True),
-    sa.Column('blog_id', sa.BigInteger(), nullable=True),
-    sa.Column('is_main', sa.Boolean(), nullable=False),
-    sa.ForeignKeyConstraint(['article_id'], ['Article.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('notification',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('notification_blogname', sa.String(length=20), nullable=True),
-    sa.Column('username', sa.String(length=20), nullable=True),
-    sa.Column('notification_type', sa.Integer(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(), server_default=sa.text('now()'), nullable=False),
-    sa.Column('checked', sa.Boolean(), server_default='0', nullable=False),
-    sa.Column('notification_blog_image_url', sa.String(length=255), nullable=True),
-    sa.Column('user_id', sa.BigInteger(), nullable=False),
-    sa.Column('blog_id', sa.BigInteger(), nullable=False),
-    sa.Column('article_id', sa.BigInteger(), nullable=True),
-    sa.Column('comment_id', sa.BigInteger(), nullable=True),
-    sa.CheckConstraint('notification_type IN (1, 2, 3, 4, 5)', name='valid_notification_type'),
-    sa.ForeignKeyConstraint(['article_id'], ['Article.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['blog_id'], ['blog.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['comment_id'], ['comment.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['user.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('notification')
-    op.drop_table('image')
-    op.drop_index(op.f('ix_comment_blog_id'), table_name='comment')
-    op.drop_index(op.f('ix_comment_article_id'), table_name='comment')
-    op.drop_table('comment')
-    op.drop_table('blog_like')
-    op.drop_index(op.f('ix_Article_title'), table_name='Article')
-    op.drop_table('Article')
-    op.drop_index(op.f('ix_draft_title'), table_name='draft')
-    op.drop_table('draft')
-    op.drop_index(op.f('ix_category_name'), table_name='category')
-    op.drop_table('category')
-    op.drop_table('blog_subscription')
-    op.drop_table('message')
-    op.drop_table('blog')
-    op.drop_index(op.f('ix_user_username'), table_name='user')
-    op.drop_index(op.f('ix_user_email'), table_name='user')
-    op.drop_table('user')
-    op.drop_table('blocked_tokens')
-    # ### end Alembic commands ###
