--- alembic/versions/a3de849b1600_initial_migration.py
@@ -1,32 +1,0 @@
-"""Initial migration
-
-Revision ID: a3de849b1600
-Revises: ed24b6c6f47e
-Create Date: 2025-03-02 13:43:44.059495
-
-"""
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-# revision identifiers, used by Alembic.
-revision = 'a3de849b1600'
-down_revision = 'ed24b6c6f47e'
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'username',
-               existing_type=mysql.VARCHAR(length=20),
-               nullable=True)
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.alter_column('user', 'username',
-               existing_type=mysql.VARCHAR(length=20),
-               nullable=False)
-    # ### end Alembic commands ###
