--- alembic/versions/e0ad72405d27_blog_api_추가.py
@@ -1,58 +1,0 @@
-"""Blog API 추가
-
-Revision ID: e0ad72405d27
-Revises: 
-Create Date: 2025-02-27 08:14:16.199790
-
-"""
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision = 'e0ad72405d27'
-down_revision = None
-branch_labels = None
-depends_on = None
-
-
-def upgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('blocked_tokens',
-    sa.Column('token_id', sa.String(length=255), nullable=False),
-    sa.Column('expired_at', sa.DateTime(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('token_id')
-    )
-    op.create_table('users',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('username', sa.String(length=20), nullable=False),
-    sa.Column('email', sa.String(length=255), nullable=False),
-    sa.Column('password', sa.String(length=100), nullable=False),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_username'), 'users', ['username'], unique=True)
-    op.create_table('blog',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('blog_name', sa.String(length=100), nullable=False),
-    sa.Column('description', sa.String(length=255), nullable=True),
-    sa.Column('created_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('updated_at', sa.DateTime(timezone=True), server_default=sa.text('now()'), nullable=False),
-    sa.Column('user_id', sa.BigInteger(), nullable=False),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
-    sa.PrimaryKeyConstraint('id'),
-    sa.UniqueConstraint('user_id', name='unique_user_blog')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('blog')
-    op.drop_index(op.f('ix_users_username'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.drop_table('users')
-    op.drop_table('blocked_tokens')
-    # ### end Alembic commands ###
