+++ mysol/app/user/store.py
@@ -10,10 +10,10 @@ from mysol.app.user.errors import UserNameAlreadyExistsError
 
 class UserStore:
     @transactional
-    async def add_user(self, username: str, user_id: str, password: str, birth: datetime) -> User:
+    async def add_user(self, username: str, login_id: str, password: str, birth: datetime) -> User:
         user = User(
             username=username,
-            user_id=user_id,
+            login_id=login_id,
             password=password,
             birth=birth
         )
